requirements.txt
fastapi
uvicorn
sqlalchemy
pg8000
google-cloud-sql-connector
pandas
google-cloud-storage
